Update to include Radical Candour table instead of image.
I've chosen to italicize the name of each category instead of adding breaks between the categories and descriptions because it looks more balanced and also because I wasn't able to quickly find a way to vertically align the markdown table text. If I find a way I'll come back and edit it.