Adds card UI, JSON file mentor management
I've updated the mentor UI, and moved the mechanic to just updating the mentor.json file to add new mentors. This is hopefully a robust way to scale it and move to things like pagination, better CSS management etc.

I also updated some of the plugins to match docusaurus core versions.